Checks for onbording and updated state using reciol
