+++ src/app/Onboarding/index.tsx
@@ -10,7 +10,6 @@ function Onboarding() {
   useEffect(() => {
     console.log(ae);
   }, []);
-  // const [storage, setStorage] = useState<string | null>(null);
   return (
     <div className="bg-gradient-to-r from-blue-400 to-emerald-400 h-screen flex flex-col justify-center items-center ">
       <div className="w-[375px] h-[600px] bg-gray-100 rounded-xl p-5 shadow-xl ">
