+++ src/app/PopupRouter.tsx
@@ -1,16 +1,23 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { useRecoilValue } from 'recoil';
-import { KeyringStoreStateEnum, lockState } from './atoms';
+import { keyringStoreState, lockState } from '../contentScript';
+import { KeyringStoreStateEnum } from './Constants';
 
 import Locked from './Locked';
 import Unlocked from './Unlocked';
 
 export default function PopupRouter() {
   const lockedState = useRecoilValue(lockState);
+  const keyring = useRecoilValue(keyringStoreState);
   return (
     <>
-      {lockedState === KeyringStoreStateEnum.Locked ? <Locked /> : <Unlocked />}
-      {/* <Locked /> */}
+      {keyring === KeyringStoreStateEnum.Loading ? (
+        'Loading'
+      ) : lockedState === KeyringStoreStateEnum.Locked ? (
+        <Locked />
+      ) : (
+        <Unlocked />
+      )}
     </>
   );
 }
